+++ spec/directions/bake_chapter_key_concepts_spec.rb
@@ -76,7 +76,13 @@ RSpec.describe Kitchen::Directions::BakeChapterKeyConcepts do
               </h2>
               <div data-type="metadata" style="display: none;">
                 <h1 data-type="document-title" itemprop="name">Key Concepts</h1>
-                <div class="authors" id="authors_copy_1">Authors</div><div class="publishers" id="publishers_copy_1">Publishers</div><div class="print-style" id="print-style_copy_1">Print Style</div><div class="permissions" id="permissions_copy_1">Permissions</div><div data-type="subject" id="subject_copy_1">Subject</div>
+                <span data-type="revised" id="revised_copy_1">Revised</span>
+                <span data-type="slug" id="slug_copy_1">Slug</span>
+                <div class="authors" id="authors_copy_1">Authors</div>
+                <div class="publishers" id="publishers_copy_1">Publishers</div>
+                <div class="print-style" id="print-style_copy_1">Print Style</div>
+                <div class="permissions" id="permissions_copy_1">Permissions</div>
+                <div data-type="subject" id="subject_copy_1">Subject</div>
               </div>
               <div class="os-key-concepts">
                 <div class="os-section-area">
@@ -131,7 +137,13 @@ RSpec.describe Kitchen::Directions::BakeChapterKeyConcepts do
               </h3>
               <div data-type="metadata" style="display: none;">
                 <h1 data-type="document-title" itemprop="name">Key Concepts</h1>
-                <div class="authors" id="authors_copy_1">Authors</div><div class="publishers" id="publishers_copy_1">Publishers</div><div class="print-style" id="print-style_copy_1">Print Style</div><div class="permissions" id="permissions_copy_1">Permissions</div><div data-type="subject" id="subject_copy_1">Subject</div>
+                <span data-type="revised" id="revised_copy_1">Revised</span>
+                <span data-type="slug" id="slug_copy_1">Slug</span>
+                <div class="authors" id="authors_copy_1">Authors</div>
+                <div class="publishers" id="publishers_copy_1">Publishers</div>
+                <div class="print-style" id="print-style_copy_1">Print Style</div>
+                <div class="permissions" id="permissions_copy_1">Permissions</div>
+                <div data-type="subject" id="subject_copy_1">Subject</div>
               </div>
               <div class="os-key-concepts">
                 <div class="os-section-area">