+++ spec/directions/bake_chapter_key_equations_spec.rb
@@ -59,7 +59,13 @@ RSpec.describe Kitchen::Directions::BakeChapterKeyEquations do
               </h2>
               <div data-type="metadata" style="display: none;">
                 <h1 data-type="document-title" itemprop="name">Key Equations</h1>
-                <div class="authors" id="authors_copy_1">Authors</div><div class="publishers" id="publishers_copy_1">Publishers</div><div class="print-style" id="print-style_copy_1">Print Style</div><div class="permissions" id="permissions_copy_1">Permissions</div><div data-type="subject" id="subject_copy_1">Subject</div>
+                <span data-type="revised" id="revised_copy_1">Revised</span>
+                <span data-type="slug" id="slug_copy_1">Slug</span>
+                <div class="authors" id="authors_copy_1">Authors</div>
+                <div class="publishers" id="publishers_copy_1">Publishers</div>
+                <div class="print-style" id="print-style_copy_1">Print Style</div>
+                <div class="permissions" id="permissions_copy_1">Permissions</div>
+                <div data-type="subject" id="subject_copy_1">Subject</div>
               </div>
               <section class="key-equations">
                 <p>Equations blah.</p>
@@ -92,7 +98,13 @@ RSpec.describe Kitchen::Directions::BakeChapterKeyEquations do
               </h3>
               <div data-type="metadata" style="display: none;">
                 <h1 data-type="document-title" itemprop="name">Key Equations</h1>
-                <div class="authors" id="authors_copy_1">Authors</div><div class="publishers" id="publishers_copy_1">Publishers</div><div class="print-style" id="print-style_copy_1">Print Style</div><div class="permissions" id="permissions_copy_1">Permissions</div><div data-type="subject" id="subject_copy_1">Subject</div>
+                <span data-type="revised" id="revised_copy_1">Revised</span>
+                <span data-type="slug" id="slug_copy_1">Slug</span>
+                <div class="authors" id="authors_copy_1">Authors</div>
+                <div class="publishers" id="publishers_copy_1">Publishers</div>
+                <div class="print-style" id="print-style_copy_1">Print Style</div>
+                <div class="permissions" id="permissions_copy_1">Permissions</div>
+                <div data-type="subject" id="subject_copy_1">Subject</div>
               </div>
               <section class="key-equations">
                 <p>Equations blah.</p>