+++ spec/directions/bake_chapter_summary_spec.rb
@@ -90,6 +90,8 @@ RSpec.describe Kitchen::Directions::BakeChapterSummary do
                 </h2>
                 <div data-type="metadata" style="display: none;">
                   <h1 data-type="document-title" itemprop="name">Summary</h1>
+                  <span data-type="revised" id="revised_copy_1">Revised</span>
+                  <span data-type="slug" id="slug_copy_1">Slug</span>
                   <div class="authors" id="authors_copy_1">Authors</div>
                   <div class="publishers" id="publishers_copy_1">Publishers</div>
                   <div class="print-style" id="print-style_copy_1">Print Style</div>
@@ -159,6 +161,8 @@ RSpec.describe Kitchen::Directions::BakeChapterSummary do
                 </h2>
                 <div data-type="metadata" style="display: none;">
                   <h1 data-type="document-title" itemprop="name">Summary</h1>
+                  <span data-type="revised" id="revised_copy_1">Revised</span>
+                  <span data-type="slug" id="slug_copy_1">Slug</span>
                   <div class="authors" id="authors_copy_1">Authors</div>
                   <div class="publishers" id="publishers_copy_1">Publishers</div>
                   <div class="print-style" id="print-style_copy_1">Print Style</div>
@@ -222,6 +226,8 @@ RSpec.describe Kitchen::Directions::BakeChapterSummary do
             </h2>
             <div data-type="metadata" style="display: none;">
               <h1 data-type="document-title" itemprop="name">Summary</h1>
+              <span data-type="revised" id="revised_copy_1">Revised</span>
+              <span data-type="slug" id="slug_copy_1">Slug</span>
               <div class="authors" id="authors_copy_1">Authors</div>
               <div class="publishers" id="publishers_copy_1">Publishers</div>
               <div class="print-style" id="print-style_copy_1">Print Style</div>