+++ spec/directions/bake_references/v2_spec.rb
@@ -14,6 +14,8 @@ RSpec.describe Kitchen::Directions::BakeReferences::V2 do
     book_containing(html:
     <<~HTML
       <div data-type="metadata" style="display: none;">
+        <span data-type="revised" id="revised_copy_1">Revised</span>
+        <span data-type="slug" id="slug_copy_1">Slug</span>
         <div class="authors" id="authors">Authors</div>
         <div class="publishers" id="publishers">Publishers</div>
         <div class="print-style" id="print-style">Print Style</div>
@@ -65,6 +67,8 @@ RSpec.describe Kitchen::Directions::BakeReferences::V2 do
           </h1>
           <div data-type="metadata" style="display: none;">
             <h1 data-type="document-title" itemprop="name">References</h1>
+            <span data-type="revised" id="revised_copy_1">Revised</span>
+            <span data-type="slug" id="slug_copy_1">Slug</span>
             <div class="authors" id="authors_copy_1">Authors</div>
             <div class="publishers" id="publishers_copy_1">Publishers</div>
             <div class="print-style" id="print-style_copy_1">Print Style</div>