+++ spec/directions/book_answer_key_container/v1_spec.rb
@@ -25,6 +25,8 @@ RSpec.describe Kitchen::Directions::BookAnswerKeyContainer::V1 do
           </h1>
           <div data-type="metadata" style="display: none;">
             <h1 data-type="document-title" itemprop="name">Answer Key</h1>
+            <span data-type="revised" id="revised_copy_1">Revised</span>
+            <span data-type="slug" id="slug_copy_1">Slug</span>
             <div class="authors" id="authors_copy_1">Authors</div>
             <div class="publishers" id="publishers_copy_1">Publishers</div>
             <div class="print-style" id="print-style_copy_1">Print Style</div>
@@ -45,6 +47,8 @@ RSpec.describe Kitchen::Directions::BookAnswerKeyContainer::V1 do
           </h1>
           <div data-type="metadata" style="display: none;">
             <h1 data-type="document-title" itemprop="name">Answer Key</h1>
+            <span data-type="revised" id="revised_copy_1">Revised</span>
+            <span data-type="slug" id="slug_copy_1">Slug</span>
             <div class="authors" id="authors_copy_1">Authors</div>
             <div class="publishers" id="publishers_copy_1">Publishers</div>
             <div class="print-style" id="print-style_copy_1">Print Style</div>