+++ spec/directions/eoc_composite_page_container/v1_spec.rb
@@ -57,6 +57,8 @@ RSpec.describe Kitchen::Directions::EocCompositePageContainer do
             </h3>
             <div data-type="metadata" style="display: none;">
               <h1 data-type="document-title" itemprop="name">Some Eoc Section</h1>
+              <span data-type="revised" id="revised_copy_1">Revised</span>
+              <span data-type="slug" id="slug_copy_1">Slug</span>
               <div class="authors" id="authors_copy_1">Authors</div>
               <div class="publishers" id="publishers_copy_1">Publishers</div>
               <div class="print-style" id="print-style_copy_1">Print Style</div>