+++ spec/directions/move_custom_section_to_eoc_container/v1_spec.rb
@@ -92,6 +92,8 @@ RSpec.describe Kitchen::Directions::MoveCustomSectionToEocContainer do
               </h3>
               <div data-type="metadata" style="display: none;">
                 <h1 data-type="document-title" itemprop="name">Some Eoc Section</h1>
+                <span data-type="revised" id="revised_copy_1">Revised</span>
+                <span data-type="slug" id="slug_copy_1">Slug</span>
                 <div class="authors" id="authors_copy_1">Authors</div>
                 <div class="publishers" id="publishers_copy_1">Publishers</div>
                 <div class="print-style" id="print-style_copy_1">Print Style</div>
@@ -131,6 +133,8 @@ RSpec.describe Kitchen::Directions::MoveCustomSectionToEocContainer do
             </h3>
             <div data-type="metadata" style="display: none;">
               <h1 data-type="document-title" itemprop="name">Some Eoc Section</h1>
+              <span data-type="revised" id="revised_copy_1">Revised</span>
+              <span data-type="slug" id="slug_copy_1">Slug</span>
               <div class="authors" id="authors_copy_1">Authors</div>
               <div class="publishers" id="publishers_copy_1">Publishers</div>
               <div class="print-style" id="print-style_copy_1">Print Style</div>
@@ -176,6 +180,8 @@ RSpec.describe Kitchen::Directions::MoveCustomSectionToEocContainer do
             </h3>
             <div data-type="metadata" style="display: none;">
               <h1 data-type="document-title" itemprop="name">Some Eoc Section</h1>
+              <span data-type="revised" id="revised_copy_1">Revised</span>
+              <span data-type="slug" id="slug_copy_1">Slug</span>
               <div class="authors" id="authors_copy_1">Authors</div>
               <div class="publishers" id="publishers_copy_1">Publishers</div>
               <div class="print-style" id="print-style_copy_1">Print Style</div>
@@ -222,6 +228,8 @@ RSpec.describe Kitchen::Directions::MoveCustomSectionToEocContainer do
             </h2>
             <div data-type="metadata" style="display: none;">
               <h1 data-type="document-title" itemprop="name">Some Eoc Section</h1>
+              <span data-type="revised" id="revised_copy_1">Revised</span>
+              <span data-type="slug" id="slug_copy_1">Slug</span>
               <div class="authors" id="authors_copy_1">Authors</div>
               <div class="publishers" id="publishers_copy_1">Publishers</div>
               <div class="print-style" id="print-style_copy_1">Print Style</div>