+++ spec/directions/move_exercises_to_eoc/v1_spec.rb
@@ -70,6 +70,8 @@ RSpec.describe Kitchen::Directions::MoveExercisesToEOC::V1 do
                   </h3>
                   <div data-type="metadata" style="display: none;">
                     <h1 data-type="document-title" itemprop="name">foo</h1>
+                    <span data-type="revised" id="revised_copy_1">Revised</span>
+                    <span data-type="slug" id="slug_copy_1">Slug</span>
                     <div class="authors" id="authors_copy_1">Authors</div>
                     <div class="publishers" id="publishers_copy_1">Publishers</div>
                     <div class="print-style" id="print-style_copy_1">Print Style</div>
@@ -108,6 +110,8 @@ RSpec.describe Kitchen::Directions::MoveExercisesToEOC::V1 do
               </h2>
               <div data-type="metadata" style="display: none;">
                 <h1 data-type="document-title" itemprop="name">foo</h1>
+                <span data-type="revised" id="revised_copy_1">Revised</span>
+                <span data-type="slug" id="slug_copy_1">Slug</span>
                 <div class="authors" id="authors_copy_1">Authors</div>
                 <div class="publishers" id="publishers_copy_1">Publishers</div>
                 <div class="print-style" id="print-style_copy_1">Print Style</div>