+++ spec/directions/move_solutions_to_answer_key/uphysics_strategy_spec.rb
@@ -122,6 +122,8 @@ RSpec.describe Kitchen::Directions::MoveSolutionsToAnswerKey::V1 do
             </h2>
             <div data-type="metadata" style="display: none;">
               <h1 data-type="document-title" itemprop="name">Chapter 1</h1>
+              <span data-type="revised" id="revised_copy_1">Revised</span>
+              <span data-type="slug" id="slug_copy_1">Slug</span>
               <div class="authors" id="authors_copy_1">Authors</div>
               <div class="publishers" id="publishers_copy_1">Publishers</div>
               <div class="print-style" id="print-style_copy_1">Print Style</div>
@@ -162,6 +164,8 @@ RSpec.describe Kitchen::Directions::MoveSolutionsToAnswerKey::V1 do
             </h2>
             <div data-type="metadata" style="display: none;">
               <h1 data-type="document-title" itemprop="name">Chapter 2</h1>
+              <span data-type="revised" id="revised_copy_1">Revised</span>
+              <span data-type="slug" id="slug_copy_1">Slug</span>
               <div class="authors" id="authors_copy_1">Authors</div>
               <div class="publishers" id="publishers_copy_1">Publishers</div>
               <div class="print-style" id="print-style_copy_1">Print Style</div>